2026PKUCourseHW5:fix_overflow

+++ source/source_relax/bfgs.cpp
@@ -128,7 +128,7 @@ void BFGS::PrepareStep(std::vector<ModuleBase::Vector3<double>>& force,
     //! call dysev
     std::vector<double> omega(3*size);
     std::vector<double> work(3*size*3*size);
-    int max_size_t=std::numeric_limits<size_t>::max();
+    size_t max_size_t=std::numeric_limits<size_t>::max();
     size_t lwork;
     if(size>=max_size_t/(3*3)/size){
        lwork=3*size*3*size;
